design bugs resolved
