+++ react-fm/src/pages/aboutUs/AboutUs.module.scss
@@ -4,6 +4,7 @@
 
 .box {
     flex-flow: column;
+    height: 95vh;
 }
 
 .ourStory {
